Update debounce value in server.py and index.html
This commit updates the debounce value in the server.py file to use the new value from the data.json file. It also updates the index.html file to display the current debounce value in the input field. These changes ensure that the debounce functionality is properly synchronized between the server and the client.